feat: torch.float8_e4m3fn compute dtype

+++ plugins/accelerated-peft/src/fms_acceleration_peft/framework_plugin_bnb.py
@@ -144,7 +144,7 @@ class BNBAccelerationPlugin(AccelerationPlugin):
             load_in_4bit=True,
             bnb_4bit_use_double_quant=True,
             bnb_4bit_quant_type=self._quant_type,
-            bnb_4bit_compute_dtype=torch_dtype,
+            bnb_4bit_compute_dtype=torch.float8_e4m3fn,
             **config_kwargs,
         )
